// frontend/src/types/exam.ts
export type MemberTransaction = {
  id: string;
  code: string;
  type: 'MEMBERSHIP' | 'ADDON';
  amount: number;
  method: string;
  status: 'PENDING' | 'PAID' | 'REJECTED';
  createdAt: string;
  proofUrl?: string | null;
  description?: string | null;
  activatedAt?: string | null;
  expiresAt?: string | null;
  package: { name: string; category: string };
  addon?: { name: string } | null;
};